+++ NEWS.md
@@ -1,6 +1,7 @@
 # staccuracy (development version)
 
-* Added [sa_diff()] function for bootstrapped-based comparison of staccuracies.
+* Added {sa_diff()} function for bootstrapped-based comparison of staccuracies.
+* var_type() is no longer exported since it is really an internal function.
 
 # staccuracy 0.1.0